+++ tests/integration/cloudwanderer/test_cloud_wanderer_write_resource.py
@@ -104,7 +104,8 @@ class TestCloudWandererWriteResource(unittest.TestCase, GenericAssertionHelpers)
                         region="us-east-1",
                         service="iam",
                         resource_type="role_policy",
-                        resource_id="test-role/test-role-policy",
+                        parent_resource_id="test-role",
+                        resource_id="test-role-policy",
                     )
                 ],
             },